Split ExecutionPlanRunner into partial files by concern
Split the large ExecutionPlanRunner class (~3100 lines) into focused partial files:

- Types.cs: Nested types (AwaitState, AsyncGeneratorStepResult, enums, TryFrame, etc.)
- Helpers.cs: Static helper methods (CreateIteratorResult, CreateIteratorDriverState, etc.)
- Completion.cs: Completion/control flow methods (RecordYield, HandleAbruptCompletion, etc.)

Main file now contains only:
- Core fields and lazy state accessors
- Constructor and public methods
- ExecutePlan main loop

This is a pure code organization refactor with no behavioral changes.
