+++ src/Asynkron.JsEngine/Ast/TypedAstEvaluator.ExecutionPlanRunner.Helpers.cs
@@ -1,0 +1,136 @@
+#region
+
+using System.Runtime.CompilerServices;
+using Asynkron.JsEngine.Execution;
+using Asynkron.JsEngine.JsTypes;
+using Asynkron.JsEngine.StdLib;
+
+#endregion
+
+namespace Asynkron.JsEngine.Ast;
+
+public static partial class TypedAstEvaluator
+{
+    private sealed partial class ExecutionPlanRunner
+    {
+        private static JsValue CreateIteratorResult(JsValue value, bool done)
+        {
+            // Use singleton for the common done case with undefined value
+            if (done && value.IsUndefined)
+            {
+                return IteratorResultObject.DoneUndefined.AsJsValue;
+            }
+
+            return new IteratorResultObject(value, done).AsJsValue;
+        }
+
+        private static IteratorDriverState CreateIteratorDriverState(
+            JsValue iterable,
+            IteratorDriverKind kind,
+            EvaluationContext context)
+        {
+            // FAST PATH: Use IEnumerator<JsValue> for arrays to avoid iterator object allocation.
+            // This bypasses creating iterator objects with next() methods for JsArray.
+            var fastEnumerator = TryGetFastEnumeratorForIteration(iterable);
+            if (fastEnumerator is not null)
+            {
+                return new IteratorDriverState
+                {
+                    IteratorObject = null,
+                    Enumerator = fastEnumerator,
+                    IsAsyncIterator = kind == IteratorDriverKind.Await,
+                    NextMethod = null
+                };
+            }
+
+            // SLOW PATH: Full iterator protocol for custom iterables
+            var iteratorTarget = NormalizeIterableTarget(iterable, context);
+
+            if (!TryGetIteratorFromProtocols(iteratorTarget, context, out var iterator) || iterator is null)
+            {
+                throw StandardLibrary.ThrowTypeError("Value is not iterable", context, context.RealmState);
+            }
+
+            var nextMethod = iterator.GetIteratorNextCallable(context);
+            return new IteratorDriverState
+            {
+                IteratorObject = iterator,
+                Enumerator = null,
+                IsAsyncIterator = kind == IteratorDriverKind.Await,
+                NextMethod = nextMethod
+            };
+        }
+
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        private static void StoreSymbolValue(JsEnvironment environment, Symbol symbol, object? /* intentional */ value)
+        {
+            // Handle case where value is already a boxed JsValue
+            var jsVal = value is JsValue jv ? jv : JsValue.FromObjectUnsafe(value);
+            StoreSymbolValueJsValue(environment, symbol, jsVal);
+        }
+
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        private static void StoreSymbolValueJsValue(JsEnvironment environment, Symbol symbol, JsValue value)
+        {
+            // DefineOrAssignJsValue is O(1) on the current environment -
+            // it only looks at environment.Values, no scope chain walk.
+            // This is optimal for generator symbols defined in the execution environment.
+            environment.DefineOrAssignJsValue(symbol, value);
+        }
+
+        /// <summary>
+        /// Stores a value using pre-resolved slot index for O(1) access.
+        /// Falls back to dictionary-based storage if slot index is invalid.
+        /// </summary>
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        private static void StoreValueBySlot(JsEnvironment environment, Symbol symbol, int slotIndex, JsValue value)
+        {
+            if (slotIndex >= 0 && environment.HasSlots)
+            {
+                environment.GetSlotRef(slotIndex) = value;
+                // Also update dictionary for symbol-based lookups elsewhere
+                environment.DefineOrAssignJsValue(symbol, value);
+            }
+            else
+            {
+                environment.DefineOrAssignJsValue(symbol, value);
+            }
+        }
+
+        /// <summary>
+        /// Reads a value using pre-resolved slot index for O(1) access.
+        /// Falls back to dictionary-based lookup if slot index is invalid.
+        /// </summary>
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        private static bool TryGetValueBySlot(JsEnvironment environment, Symbol symbol, int slotIndex, out JsValue value)
+        {
+            if (slotIndex >= 0 && environment.HasSlots)
+            {
+                value = environment.GetSlotRef(slotIndex);
+                return true;
+            }
+
+            return TryGetSymbolValueJsValue(environment, symbol, out value);
+        }
+
+        private static bool TryGetSymbolValueJsValue(JsEnvironment environment, Symbol symbol, out JsValue value)
+        {
+            if (environment.TryGetJsValue(symbol, out value))
+            {
+                return true;
+            }
+
+            value = JsValue.Undefined;
+            return false;
+        }
+
+        private static JsValue FinishExternalCompletion(ResumeMode mode, JsValue value)
+        {
+            return mode switch
+            {
+                ResumeMode.Throw => throw new ThrowSignal(value),
+                _ => CreateIteratorResult(value, true)
+            };
+        }
+    }
+}
