+++ src/Asynkron.JsEngine/Ast/TypedAstEvaluator.ExecutionPlanRunner.cs
@@ -390,114 +390,6 @@ public static partial class TypedAstEvaluator
 
             return executionEnvironment;
         }
-        private static JsValue CreateIteratorResult(JsValue value, bool done)
-        {
-            // Use singleton for the common done case with undefined value
-            if (done && value.IsUndefined)
-            {
-                return IteratorResultObject.DoneUndefined.AsJsValue;
-            }
-            return new IteratorResultObject(value, done).AsJsValue;
-        }
-
-        private static IteratorDriverState CreateIteratorDriverState(
-            JsValue iterable,
-            IteratorDriverKind kind,
-            EvaluationContext context)
-        {
-            // FAST PATH: Use IEnumerator<JsValue> for arrays to avoid iterator object allocation.
-            // This bypasses creating iterator objects with next() methods for JsArray.
-            var fastEnumerator = TryGetFastEnumeratorForIteration(iterable);
-            if (fastEnumerator is not null)
-            {
-                return new IteratorDriverState
-                {
-                    IteratorObject = null,
-                    Enumerator = fastEnumerator,
-                    IsAsyncIterator = kind == IteratorDriverKind.Await,
-                    NextMethod = null
-                };
-            }
-
-            // SLOW PATH: Full iterator protocol for custom iterables
-            var iteratorTarget = NormalizeIterableTarget(iterable, context);
-
-            if (!TryGetIteratorFromProtocols(iteratorTarget, context, out var iterator) || iterator is null)
-            {
-                throw StandardLibrary.ThrowTypeError("Value is not iterable", context, context.RealmState);
-            }
-
-            var nextMethod = iterator.GetIteratorNextCallable(context);
-            return new IteratorDriverState
-            {
-                IteratorObject = iterator,
-                Enumerator = null,
-                IsAsyncIterator = kind == IteratorDriverKind.Await,
-                NextMethod = nextMethod
-            };
-        }
-
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        private static void StoreSymbolValue(JsEnvironment environment, Symbol symbol, object? /* intentional */ value)
-        {
-            // Handle case where value is already a boxed JsValue
-            var jsVal = value is JsValue jv ? jv : JsValue.FromObjectUnsafe(value);
-            StoreSymbolValueJsValue(environment, symbol, jsVal);
-        }
-
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        private static void StoreSymbolValueJsValue(JsEnvironment environment, Symbol symbol, JsValue value)
-        {
-            // DefineOrAssignJsValue is O(1) on the current environment -
-            // it only looks at environment.Values, no scope chain walk.
-            // This is optimal for generator symbols defined in the execution environment.
-            environment.DefineOrAssignJsValue(symbol, value);
-        }
-
-        /// <summary>
-        /// Stores a value using pre-resolved slot index for O(1) access.
-        /// Falls back to dictionary-based storage if slot index is invalid.
-        /// </summary>
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        private static void StoreValueBySlot(JsEnvironment environment, Symbol symbol, int slotIndex, JsValue value)
-        {
-            if (slotIndex >= 0 && environment.HasSlots)
-            {
-                environment.GetSlotRef(slotIndex) = value;
-                // Also update dictionary for symbol-based lookups elsewhere
-                environment.DefineOrAssignJsValue(symbol, value);
-            }
-            else
-            {
-                environment.DefineOrAssignJsValue(symbol, value);
-            }
-        }
-
-        /// <summary>
-        /// Reads a value using pre-resolved slot index for O(1) access.
-        /// Falls back to dictionary-based lookup if slot index is invalid.
-        /// </summary>
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        private static bool TryGetValueBySlot(JsEnvironment environment, Symbol symbol, int slotIndex, out JsValue value)
-        {
-            if (slotIndex >= 0 && environment.HasSlots)
-            {
-                value = environment.GetSlotRef(slotIndex);
-                return true;
-            }
-            return TryGetSymbolValueJsValue(environment, symbol, out value);
-        }
-
-        private static bool TryGetSymbolValueJsValue(JsEnvironment environment, Symbol symbol, out JsValue value)
-        {
-            if (environment.TryGetJsValue(symbol, out value))
-            {
-                return true;
-            }
-
-            value = JsValue.Undefined;
-            return false;
-        }
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         private JsValue ExecutePlan(ResumeMode mode, JsValue resumeValue)
@@ -2548,15 +2440,6 @@ public static partial class TypedAstEvaluator
             return _isStrict ? ScopeMode.Strict : ScopeMode.Sloppy;
         }
 
-        private static JsValue FinishExternalCompletion(ResumeMode mode, JsValue value)
-        {
-            return mode switch
-            {
-                ResumeMode.Throw => throw new ThrowSignal(value),
-                _ => CreateIteratorResult(value, true)
-            };
-        }
-
         internal JsValue EvaluateAwaitInGenerator(AwaitExpression expression, JsEnvironment environment,
             EvaluationContext context)
         {
@@ -2689,432 +2572,5 @@ public static partial class TypedAstEvaluator
             return true;
         }
 
-        private void RecordYield(EvaluationContext context, JsEnvironment? currentEnvironment = null)
-        {
-            // Remember the active yield slot so the next resume value is applied to the
-            // right YieldExpression (ECMA-262 GeneratorResume, step threading of sent values).
-            YieldStateRef.LastYieldIndex = context.LastYieldIndex;
-
-            // Also save source positions for yields from StatementInstruction (AST-evaluated yields).
-            // These are used to set up resume state so the yield expression returns the resume value.
-            YieldStateRef.LastYieldSourceStart = context.LastYieldSourceStart;
-            YieldStateRef.LastYieldSourceEnd = context.LastYieldSourceEnd;
-
-            _realmState.Logger?.LogInformation(
-                "RecordYield: yieldIndex={YieldIndex} sourceStart={Start} sourceEnd={End}",
-                YieldStateRef.LastYieldIndex, YieldStateRef.LastYieldSourceStart, YieldStateRef.LastYieldSourceEnd);
-
-            // Save the current environment so that when the generator resumes, it uses
-            // the correct per-iteration environment (for loops with let bindings).
-            // This is critical for `continue` to work correctly in generators.
-            if (currentEnvironment != null)
-            {
-                _executionEnvironment = currentEnvironment;
-            }
-
-            // Clear the context's source positions for the next yield
-            context.LastYieldSourceStart = -1;
-            context.LastYieldSourceEnd = -1;
-        }
-
-
-        private void PreparePendingResumeValue(ResumeMode mode, JsValue resumeValue, bool wasStart)
-        {
-            if (wasStart)
-            {
-                // Per ES spec: The first next() argument is ignored when starting a generator.
-                // This applies to both regular yield and yield* - the first call to inner iterator's
-                // next() receives undefined, not the outer generator's first next() argument.
-                AsyncStateRef.PendingResumeKind = ResumePayloadKind.None;
-                AsyncStateRef.PendingResumeValue = JsValue.Undefined;
-                return;
-            }
-
-            AsyncStateRef.PendingResumeKind = mode switch
-            {
-                ResumeMode.Throw => ResumePayloadKind.Throw,
-                ResumeMode.Return => ResumePayloadKind.Return,
-                _ => ResumePayloadKind.Value
-            };
-
-            AsyncStateRef.PendingResumeValue = resumeValue;
-
-            if (_realmState.Logger?.IsEnabled(LogLevel.Information) == true)
-            {
-                var resumeType = resumeValue.ObjectValue?.GetType().Name ?? resumeValue.Kind.ToString();
-                _realmState.Logger.LogInformation(
-                    "PrepareResume yieldIndex={YieldIndex} kind={Kind} valueType={Type}",
-                    YieldStateRef.LastYieldIndex,
-                    AsyncStateRef.PendingResumeKind,
-                    resumeType);
-            }
-
-            if (YieldStateRef.LastYieldIndex < 0)
-            {
-                return;
-            }
-
-            var resumeSlotIndex = YieldStateRef.LastYieldIndex;
-            switch (AsyncStateRef.PendingResumeKind)
-            {
-                case ResumePayloadKind.Throw:
-                    YieldStateRef.ResumeContext.SetException(resumeSlotIndex, resumeValue);
-                    break;
-                case ResumePayloadKind.Return:
-                    YieldStateRef.ResumeContext.SetReturn(resumeSlotIndex, resumeValue);
-                    break;
-                default:
-                    YieldStateRef.ResumeContext.SetValue(resumeSlotIndex, resumeValue);
-                    break;
-            }
-        }
-
-        private (ResumePayloadKind Kind, JsValue Value) ConsumeResumeValue()
-        {
-            var kind = AsyncStateRef.PendingResumeKind;
-            var value = AsyncStateRef.PendingResumeValue;
-            AsyncStateRef.PendingResumeKind = ResumePayloadKind.None;
-            AsyncStateRef.PendingResumeValue = JsValue.Undefined;
-
-            if (kind == ResumePayloadKind.None)
-            {
-                return (ResumePayloadKind.Value, JsValue.Undefined);
-            }
-
-            return (kind, value);
-        }
-
-        private void PushTryFrame(EnterTryInstruction instruction, JsEnvironment environment)
-        {
-            var frame = new TryFrame(instruction.HandlerIndex, instruction.CatchSlotSymbol, instruction.FinallyIndex, environment);
-            if (instruction.CatchSlotSymbol is { } slot && !environment.HasBinding(slot))
-            {
-                environment.DefineJsValue(slot, JsValue.Undefined);
-            }
-
-            TryCatchStateRef.TryStack.Push(frame);
-        }
-
-        /// <summary>
-        /// Finds the jump target for a break or continue signal.
-        /// For unlabeled: returns the innermost loop's target.
-        /// For labeled: searches the stack for a matching label.
-        /// </summary>
-        private int FindLoopTarget(Symbol? label, bool isBreak)
-        {
-            if (LoopStateRef.LoopStack.Count == 0)
-            {
-                return -1;
-            }
-
-            if (label is null)
-            {
-                // Unlabeled: use innermost loop
-                var frame = LoopStateRef.LoopStack.Peek();
-                return isBreak ? frame.BreakTarget : frame.ContinueTarget;
-            }
-
-            // Labeled: search for matching label
-            foreach (var frame in LoopStateRef.LoopStack)
-            {
-                if (ReferenceEquals(frame.Label, label))
-                {
-                    return isBreak ? frame.BreakTarget : frame.ContinueTarget;
-                }
-            }
-
-            return -1;
-        }
-
-        private void CompleteTryNormally(int resumeTarget)
-        {
-            if (TryCatchStateRef.TryStack.Count == 0)
-            {
-                _programCounter = resumeTarget;
-                return;
-            }
-
-            var frame = TryCatchStateRef.TryStack.Peek();
-            if (frame is { FinallyIndex: >= 0, FinallyScheduled: false })
-            {
-                frame.FinallyScheduled = true;
-                frame.PendingCompletion = PendingCompletion.FromNormal(resumeTarget);
-                _programCounter = frame.FinallyIndex;
-                return;
-            }
-
-            TryCatchStateRef.TryStack.Pop();
-            _programCounter = resumeTarget;
-        }
-
-        private bool HandleAbruptCompletion(AbruptKind kind, object? /* intentional */ value, JsEnvironment environment)
-        {
-            // Clear any previous restored environment
-            TryCatchStateRef.RestoredEnvironmentFromTry = null;
-
-            while (TryCatchStateRef.TryStack.Count > 0)
-            {
-                var frame = TryCatchStateRef.TryStack.Peek();
-                if (kind == AbruptKind.Throw && frame is { HandlerIndex: >= 0, CatchUsed: false })
-                {
-                    frame.CatchUsed = true;
-
-                    // Restore to the environment that was active when entering the try block.
-                    // This ensures that block-scoped bindings inside the try are no longer visible.
-                    var targetEnv = frame.EntryEnvironment;
-                    TryCatchStateRef.RestoredEnvironmentFromTry = targetEnv;
-
-                    if (frame.CatchSlotSymbol is { } slot)
-                    {
-                        // Handle case where value is already a boxed JsValue
-                        var valueJs = value is JsValue js ? js : JsValue.FromObjectUnsafe(value);
-                        // Use the entry environment for the catch slot, not the current (nested) environment
-                        if (targetEnv.HasBinding(slot))
-                        {
-                            targetEnv.AssignJsValue(slot, valueJs);
-                        }
-                        else
-                        {
-                            targetEnv.DefineJsValue(slot, valueJs);
-                        }
-                    }
-
-                    _programCounter = frame.HandlerIndex;
-                    return true;
-                }
-
-                if (frame.FinallyIndex >= 0)
-                {
-                    if (!frame.FinallyScheduled)
-                    {
-                        frame.FinallyScheduled = true;
-                        frame.PendingCompletion = PendingCompletion.FromAbrupt(kind, value);
-
-                        // Restore to the environment that was active when entering the try block.
-                        // This ensures that block-scoped bindings inside the try are no longer visible in finally.
-                        TryCatchStateRef.RestoredEnvironmentFromTry = frame.EntryEnvironment;
-
-                        _programCounter = frame.FinallyIndex;
-                        return true;
-                    }
-
-                    // Per ES spec: when an abrupt completion occurs inside a finally block,
-                    // the new completion replaces the pending one. For most callers
-                    // (like StoreResumeValueInstruction for generator resumption),
-                    // we update the pending and let them advance PC. For throw/return
-                    // statements that end the finally abruptly, the caller handles it.
-                    frame.PendingCompletion = PendingCompletion.FromAbrupt(kind, value);
-                    return true;
-                }
-
-                TryCatchStateRef.TryStack.Pop();
-            }
-
-            return false;
-        }
-
-        /// <summary>
-        /// JsValue overload - boxes the JsValue which is better than ToObject() as it preserves type info.
-        /// </summary>
-        private bool HandleAbruptCompletionJsValue(AbruptKind kind, JsValue value, JsEnvironment environment)
-        {
-            // Boxing JsValue is preferred over ToObject() because:
-            // 1. It preserves the JsValue type information
-            // 2. Downstream code can detect "is JsValue" and unbox efficiently
-            return HandleAbruptCompletion(kind, value, environment);
-        }
-
-        private JsValue CompleteReturn(JsValue value)
-        {
-            // Close any active iterators before completing (array patterns and for-of loops)
-            CloseActiveIterators();
-
-            _programCounter = -1;
-            _state = GeneratorState.Completed;
-            _done = true;
-            TryCatchStateRef.TryStack.Clear();
-            return CreateIteratorResult(value, true);
-        }
-
-        /// <summary>
-        /// Closes all active iterators stored in the environment chain.
-        /// This handles both array pattern iterators (from destructuring) and
-        /// for-of loop iterators via the unified IActiveIteratorState interface.
-        /// </summary>
-        private void CloseActiveIterators()
-        {
-            if (_executionEnvironment is null)
-            {
-                return;
-            }
-
-            // Create a fresh context to avoid state interference from the existing context
-            var context = _realmState.CreateContext(ScopeKind.Function, DetermineGeneratorScopeMode());
-
-            // Scan the environment for all IActiveIteratorState objects and close their iterators
-            var statesToClose = new List<(IActiveIteratorState State, IJsObjectLike Iterator)>();
-            ScanEnvironmentForActiveIterators(_executionEnvironment, statesToClose);
-
-            foreach (var (state, iterator) in statesToClose)
-            {
-                // Mark as closed first (before potential exception)
-                state.MarkIteratorClosed();
-
-                // Close the iterator - if it throws, that error replaces the return completion
-                // Let the exception propagate directly without catching
-                iterator.IteratorClose(context);
-            }
-        }
-
-        /// <summary>
-        /// Scans the environment chain for all IActiveIteratorState objects and collects
-        /// those with active iterators that need closing.
-        /// </summary>
-        private static void ScanEnvironmentForActiveIterators(JsEnvironment env,
-            List<(IActiveIteratorState, IJsObjectLike)> results)
-        {
-            while (true)
-            {
-                // Scan symbol bindings in this environment (using safe method to skip uninitialized TDZ bindings)
-                foreach (var symbol in env.GetBindingSymbols())
-                {
-                    if (env.TryGetJsValueLocalSafe(symbol, out var jsValue) &&
-                        !jsValue.IsNullOrUndefined &&
-                        jsValue.TryGetObject<IActiveIteratorState>(out var state) &&
-                        state.TryGetActiveIterator(out var iterator))
-                    {
-                        results.Add((state, iterator));
-                    }
-                }
-
-                // Also scan slots for IteratorDriverState (for-of loop state)
-                if (env.HasSlots && env._slots is { } slots)
-                {
-                    foreach (var slot in slots)
-                    {
-                        if (!slot.IsNullOrUndefined &&
-                            slot.TryGetObject<IActiveIteratorState>(out var state) &&
-                            state.TryGetActiveIterator(out var iterator))
-                        {
-                            results.Add((state, iterator));
-                        }
-                    }
-                }
-
-                // Also scan parent environments
-                if (env.Enclosing is { } parent)
-                {
-                    env = parent;
-                    continue;
-                }
-
-                break;
-            }
-        }
-
-        private sealed class AwaitState
-        {
-            public bool HasResult { get; set; }
-            public bool IsThrow { get; set; }
-            public JsValue Result { get; set; } = JsValue.Undefined;
-        }
-
-        // Lightweight step result used by async-generator wrappers so they can
-        // drive the same IR plan without duplicating the interpreter. This
-        // supports yield/completion/throw, and has room for a future "Pending"
-        // state that surfaces promise-like values without blocking.
-        [StructLayout(LayoutKind.Auto)]
-        internal readonly record struct AsyncGeneratorStepResult(
-            AsyncGeneratorStepKind Kind,
-            JsValue Value,
-            bool Done,
-            JsValue PendingPromise);
-
-        internal enum AsyncGeneratorStepKind
-        {
-            Yield,
-            Completed,
-            Throw,
-            Pending
-        }
-
-        internal enum ResumeMode
-        {
-            Next,
-            Throw,
-            Return
-        }
-
-        private enum GeneratorState
-        {
-            Start,
-            Suspended,
-            Executing,
-            Completed
-        }
-
-        private enum ResumePayloadKind
-        {
-            None,
-            Value,
-            Throw,
-            Return
-        }
-
-        private enum AbruptKind
-        {
-            None,
-            Return,
-            Throw,
-            Break,
-            Continue
-        }
-
-        private sealed class TryFrame(int handlerIndex, Symbol? catchSlotSymbol, int finallyIndex, JsEnvironment entryEnvironment)
-        {
-            public int HandlerIndex { get; } = handlerIndex;
-            public Symbol? CatchSlotSymbol { get; } = catchSlotSymbol;
-            public int FinallyIndex { get; } = finallyIndex;
-            /// <summary>
-            /// The environment that was active when entering the try block.
-            /// Used to restore scope when jumping to catch/finally handlers.
-            /// </summary>
-            public JsEnvironment EntryEnvironment { get; } = entryEnvironment;
-            public bool CatchUsed { get; set; }
-            public bool FinallyScheduled { get; set; }
-            public PendingCompletion PendingCompletion { get; set; } = PendingCompletion.None;
-        }
-
-        private readonly record struct PendingCompletion(AbruptKind Kind, object? Value, int ResumeTarget)
-        {
-            public static PendingCompletion None { get; } = new(AbruptKind.None, null, -1);
-
-            public static PendingCompletion FromNormal(int resumeTarget)
-            {
-                return new PendingCompletion(AbruptKind.None, null, resumeTarget);
-            }
-
-            public static PendingCompletion FromAbrupt(AbruptKind kind, object? value)
-            {
-                return new PendingCompletion(kind, value, -1);
-            }
-        }
-
-        /// <summary>
-        /// Tracks loop context at runtime so break/continue from AST-evaluated code
-        /// (via StatementInstruction) can resolve their jump targets.
-        /// </summary>
-        private readonly record struct LoopFrame(Symbol? Label, int BreakTarget, int ContinueTarget);
-
-        private sealed class YieldStarState
-        {
-            public DelegatedYieldState? State { get; set; }
-            public bool AwaitingResume { get; set; }
-
-            public AbruptKind PendingAbrupt { get; set; }
-
-            // Use JsValue instead of object? to avoid boxing
-            public JsValue PendingValue { get; set; }
-        }
     }
 }
